requirements.txt
flask
mysql-connector-python
google-cloud-pubsub
